+++ src/vm/sbin/vmadm.js
@@ -21,7 +21,7 @@
  *
  * CDDL HEADER END
  *
- * Copyright (c) 2018, Joyent, Inc. All rights reserved.
+ * Copyright (c) 2019, Joyent, Inc.
  *
  */
 
@@ -52,6 +52,7 @@ var COMMANDS = [
     'create-snapshot',
     'delete', 'destroy',
     'delete-snapshot',
+    'events',
     'stop', 'halt',
     'help',
     'info',
@@ -92,7 +93,10 @@ var LIST_FIELDS = {
     exit_status: {header: 'EXIT', width: 4},
     exit_timestamp: {header: 'EXIT_TIMESTAMP', width: 24},
     firewall_enabled: {header: 'FIREWALL_ENABLED', width: 16},
+    flexible_disk_size: {header: 'FLEX_DISK_SIZE', width: 14},
+    free_space: {header: 'FREE_SPACE', width: 14},
     hostname: {header: 'HOSTNAME', width: 32},
+    hvm: {header: 'HVM', width: 5},
     image_uuid: {header: 'IMAGE_UUID', width: 36},
     indestructible_delegated: {header: 'INDESTR_DATA', width: 12},
     indestructible_zoneroot: {header: 'INDESTR_ROOT', width: 12},
@@ -148,6 +152,7 @@ function usage(message, code)
     out('console <uuid>');
     out('delete <uuid>');
     out('delete-snapshot <uuid> <snapname>');
+    out('events [-fjr] [uuid]');
     out('get <uuid>');
     out('info <uuid> [type,...]');
     out('install <uuid>');
@@ -243,7 +248,7 @@ function getListProperties(field)
     return undefined;
 }
 
-function getUUID(command, p)
+function getUUID(command, p, opts)
 {
     var uuid;
 
@@ -255,6 +260,10 @@ function getUUID(command, p)
 
             return uuid;
         }
+    }
+
+    if (opts && opts.optional) {
+        return null;
     }
 
     return usage('Invalid or missing UUID for ' + command);
@@ -351,17 +360,10 @@ function parseStartArgs(args)
 function parseInfoArgs(args)
 {
     var arg;
-    var type;
     var types = [];
 
     for (arg in args) {
         types = types.concat(args[arg].split(','));
-    }
-
-    for (type in types) {
-        if (VM.INFO_TYPES.indexOf(types[type]) === -1) {
-            usage('Invalid info type: ' + types[type]);
-        }
     }
 
     if (types.length === 0) {
@@ -394,6 +396,14 @@ function addCommandOptions(command, opts, shorts)
     case 'start':
     case 'sysrq':
         // these only take uuid or 'special' args like start order=cd
+        break;
+    case 'events':
+        opts.full = Boolean;
+        shorts.f = ['--full'];
+        opts.json = Boolean;
+        shorts.j = ['--json'];
+        opts.ready = Boolean;
+        shorts.r = ['--ready'];
         break;
     case 'lookup':
         opts.json = Boolean;
@@ -714,6 +724,93 @@ function readFile(filename, callback)
     });
 }
 
+function do_events(parsed, callback) {
+    var uuid = getUUID('events', parsed, {optional: true});
+
+    var opts = {
+        name: 'vmadm CLI'
+    };
+    if (process.env.VMADM_IDENT) {
+        opts.name += util.format(' - %s', process.env.VMADM_IDENT);
+    }
+    if (uuid) {
+        opts.zonename = uuid;
+    }
+
+    VM.events(opts, vmEventHandler, vmEventReady);
+
+    // Called when a vminfod event is seen
+    function vmEventHandler(ev) {
+        if (parsed.json) {
+            console.log(JSON.stringify(ev));
+            return;
+        }
+
+        var zn = formatZonename(ev.zonename);
+        var date = formatDate(ev.date);
+
+        var alias = (ev.vm || {}).alias || '-';
+        if (alias.length > 30) {
+            alias = util.format('%s...', alias.substr(0, 27));
+        }
+
+        // format the output nicely
+        var base = util.format('[%s] %s %s %s',
+            date, zn, alias, ev.type);
+
+        delete ev.vm;
+        if (ev.changes) {
+            ev.changes.forEach(function (change) {
+                console.log('%s: %s %s :: %j -> %j',
+                    base,
+                    change.prettyPath,
+                    change.action,
+                    change.oldValue,
+                    change.newValue);
+            });
+        } else {
+            console.log(base);
+        }
+    }
+
+    // Called when the vminfod stream is ready
+    function vmEventReady(err, obj) {
+        if (err) {
+            callback(err);
+            return;
+        }
+
+        if (!parsed.ready)
+            return;
+
+        var ev = obj.ev;
+        var date = formatDate(ev.date);
+        if (parsed.json) {
+            console.log(JSON.stringify(ev));
+        } else if (parsed.full) {
+            console.log('[%s] %s (uuid %s)', date, ev.type, ev.uuid);
+        } else {
+            console.log('[%s] %s', date, ev.type);
+        }
+    }
+
+    function formatDate(date) {
+        if (parsed.full) {
+            return date.toISOString();
+        } else {
+            return date.toISOString().split('T')[1];
+        }
+    }
+
+    function formatZonename(zonename) {
+        if (parsed.full) {
+            return zonename;
+        } else {
+            return zonename.split('-')[0];
+        }
+    }
+}
+
 function main(callback)
 {
     var args = process.argv.slice(1);
@@ -754,6 +851,9 @@ function main(callback)
     VM.loglevel = 'debug';
 
     switch (command) {
+    case 'events':
+        do_events(parsed, callback);
+        break;
     case 'start':
     case 'boot':
         uuid = getUUID(command, parsed);
@@ -961,6 +1061,17 @@ function main(callback)
     case 'json':
         uuid = getUUID(command, parsed);
         VM.load(uuid, function (err, obj) {
+            if (err && err.code === 'ENOENT' && err.statusCode === 404) {
+                /*
+                 * In the vminfod world, a VM not found error is expressed a
+                 * 404 instead of a failure message directly from zoneadm(1M).
+                 * Certain tools and scripts (`node-vmadm` for example) rely on
+                 * this specific output, so we make vmadm(1M) mimic that here.
+                 */
+                err.message = util.format(
+                    'zoneadm: %s: No such zone configured', uuid);
+            }
+
             if (err) {
                 callback(err);
             } else {